Extract TestSpiderBase.
